Move the exception to the plugin so comsumers only have to import one thing
