+++ lms/services/grouping/_plugin/__init__.py
@@ -1,3 +1,4 @@
+from enum import Enum
 from typing import List, Optional
 
 from lms.models import Course, Grouping
@@ -60,3 +61,22 @@ class GroupingServicePlugin:  # pragma: no cover
         """Get the groups for a learner when they are being graded."""
 
         return None
+
+
+class GroupError(Exception):
+    """Exceptions raised by plugins."""
+
+    class ErrorCodes(str, Enum):
+        """Error codes that the FE is going to check for."""
+
+        CANVAS_STUDENT_NOT_IN_GROUP = "canvas_student_not_in_group"
+        CANVAS_GROUP_SET_EMPTY = "canvas_group_set_empty"
+        CANVAS_GROUP_SET_NOT_FOUND = "canvas_group_set_not_found"
+        BLACKBOARD_GROUP_SET_NOT_FOUND = "blackboard_group_set_not_found"
+        BLACKBOARD_GROUP_SET_EMPTY = "blackboard_group_set_empty"
+        BLACKBOARD_STUDENT_NOT_IN_GROUP = "blackboard_student_not_in_group"
+
+    def __init__(self, error_code, group_set):
+        self.error_code = error_code
+        self.details = {"group_set": group_set}
+        super().__init__(self.details)
