+++ tests/unit/lms/services/grouping/_plugin/blackboard_test.py
@@ -4,8 +4,8 @@ import pytest
 
 from lms.models import Grouping
 from lms.services.exceptions import ExternalRequestError
+from lms.services.grouping._plugin import GroupError
 from lms.services.grouping._plugin.blackboard import BlackboardGroupingPlugin
-from lms.services.grouping._plugin.exceptions import GroupError
 from tests import factories
